Minor mapreduce improvements (#303)

+++ src/device/intrinsics/simd.jl
@@ -86,8 +86,8 @@ Returns `a * b + c`.
 
 ## SIMD Shuffle Up/Down
 
-simd_shuffle_map = ((Float32, "f16"),
-                    (Float16, "f32"),
+simd_shuffle_map = ((Float32, "f32"),
+                    (Float16, "f16"),
                     (Int32,   "s.i32"),
                     (UInt32,  "u.i32"),
                     (Int16,   "s.i16"),
